Fix shared-data-cache race condition and bf16-to-numpy crashes
stage_1_prepare_data.py: DATA_CACHE_DIR is shared across every seed/
model/ablation chain (WORK_DIR has no seed or model suffix) and was
unconditionally regenerated with a plain torch.save() on every run.
Concurrent job_1 executions across chains can corrupt each other's
writes to train_data.pt, surfacing downstream as "PytorchStreamReader
failed reading file data.pkl ... internal miniz error" in job_2. Fix:
save via a per-process temp file + os.replace() (atomic on POSIX), and
skip regeneration entirely once any chain's completion flag exists.

stage_8_order_preservation.py, stage_6_hotflip_attacks.py: numpy has
no bfloat16 dtype; raw hidden-state / embedding-score tensors under
the t5-base/t5-large bf16 preset threw "TypeError: Got unsupported
ScalarType BFloat16" on .cpu().numpy(). Cast to .float() first. Audited
every .numpy() call in both hpc_version/scripts and
foundation_llm_experiments/scripts for the same pattern; these were
the only two unguarded instances of a floating-point (as opposed to
integer token-id) tensor reaching numpy.

+++ hpc_version/scripts/stage_6_hotflip_attacks.py
@@ -163,8 +163,11 @@ class HotFlipT5Attack:
         
         # Get top-k tokens
         top_values, top_indices = torch.topk(scores, top_k)
-        
-        return top_indices.cpu().numpy(), top_values.cpu().numpy()
+
+        # .float(): embedding_matrix (and so scores/top_values) is bf16 for
+        # the size-tier presets (t5-base/t5-large); numpy has no bfloat16
+        # dtype. top_indices are integer, unaffected.
+        return top_indices.cpu().numpy(), top_values.float().cpu().numpy()
     
     def attack_single(self, text, summary, num_flips=5, beam_size=10, return_ids=False):
         """
